ISSUE #3703 - split ticket helper file
